+++ CHANGELOG.md
@@ -16,6 +16,8 @@ All notable changes to the "metago" extension will be documented in this file.
    - [x] target chars is only used for narrow down searching range, not for navagition. solve the problem of typing muti target chars together may edit document by mistake
    - [x] ripple search support, type location-chars to triger far from center: one char nearby; two char current doc; three chars all opened editors
    
+
+   - [ ] bug: metaJump: '/' key could not be used as following sequence key, if already used as decorator hide key. we need to use Ctrl key as the decorator hiding key.
    - [ ] metaJump: contributed command: metaGo.Cancel doesn't exist.
    - [ ] Line selection up/down command: ?
       make Ctrl+L into a line comand trigger: if the user press i(means up, configurable), triggers the selecting up subcomand, in this subcomand the user can press i again or press k (mean down) several times. Ctrl+L, k is select down subcomand triger. any other key other than i,k would escape this comand
@@ -31,7 +33,6 @@ All notable changes to the "metago" extension will be documented in this file.
 #### Fixed
 ---
 
-
 ## Current Version
 ---
 ### V3.5.0 - April 16, 2020
@@ -39,6 +40,7 @@ All notable changes to the "metago" extension will be documented in this file.
    - [x] metaJump: ripple support, type location-chars to triger far from center: one char current paragraph(seperated by empty lines); two chars current doc; three chars all opened editors; for one and two target chars, one char decorators will pass through boundaries if possible(i.e. for one target char, no two chars decorators are needed for all candidates in the current paragraph)
    - [x] metaJump: target chars is only used for narrow down searching range, not for navagition. solve the problem of typing muti target chars together may edit document by mistake
    - [x] metaGo: what's new page to show when major and minor upgrade
+   - [x] dev: webpack support bundle third party packates
 
 #### Removed
 
@@ -47,7 +49,7 @@ All notable changes to the "metago" extension will be documented in this file.
    - [x] line selection: metaGo.selectLineUp default shortcut changed from ctrl+shift+l to ctrl+i, to avoid collision with default command of ctrl+shift+l.
 #### Fixed
    - [x] bookmark: type in popup box, would jump automaticly, may edit when type fast. should only do filter!
-
+   - [x] metaJump: fix cancel exception: after tigger jump, then press Esc would throw exception.
 ---
 ### V3.4.3 - Mar 19, 2020
 #### Added